the return of the backend
fight cors with server
refactor so ts
added loading to search
