+++ backend/index.ts
@@ -1,0 +1,20 @@
+import express, { Request, Response } from "express";
+import cors from "cors";
+import dotenv from "dotenv";
+import searchCityRoutes from "./routes/searchCityRoutes";
+
+dotenv.config();
+
+const app = express();
+
+const PORT = process.env.PORT || 3000;
+
+app.use(cors());
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
+app.use("/api/cities", searchCityRoutes);
+
+app.listen(PORT, () => {
+  console.log(`Server is running at http://localhost:${PORT}`);
+});
